Refactor AudioConcatProcessor
